+++ google.js
@@ -2034,6 +2034,6 @@ var google = [
 ]
 
 var googleClose=google.map(function(date){
-  return parseInt(date["FIELD2"],10)
+  return parseFloat(date["FIELD2"],10)
 })
 googleClose = googleClose.slice(1);
